Make JwkSurrogate private
