+++ sdk/src/main/java/com/microsoft/did/sdk/crypto/protocols/jose/jws/serialization/JwkSerializer.kt
@@ -8,6 +8,7 @@ import com.nimbusds.jose.jwk.OctetKeyPair
 import com.nimbusds.jose.jwk.OctetSequenceKey
 import com.nimbusds.jose.jwk.RSAKey
 import kotlinx.serialization.KSerializer
+import kotlinx.serialization.Serializable
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
@@ -53,5 +54,37 @@ object JwkSerializer : KSerializer<JWK> {
     override fun deserialize(decoder: Decoder): JWK {
         val jsonWebKey = decoder.decodeSerializableValue(JwkSurrogate.serializer())
         return JWK.parse(Json.encodeToString(JwkSurrogate.serializer(), jsonWebKey))
+    }
+
+    @Serializable
+    private data class JwkSurrogate(
+        // The following fields are defined in Section 3.1 of JSON Web Key
+        var kty: String = "",
+        var kid: String? = null,
+        var use: String? = null,
+        var key_ops: List<String>? = null,
+        var alg: String? = null,
+
+        // The following fields are defined in JSON Web Key Parameters Registration
+        var ext: Boolean? = null,
+
+        // The following fields are defined in Section 6 of JSON Web Algorithms
+        var crv: String? = null,
+        var x: String? = null,
+        var y: String? = null,
+        var d: String? = null,
+        var n: String? = null,
+        var e: String? = null,
+        var p: String? = null,
+        var q: String? = null,
+        var dp: String? = null,
+        var dq: String? = null,
+        var qi: String? = null,
+        var oth: List<RsaOtherPrimesInfo>? = null,
+        var k: String? = null
+    ) {
+        /** The following fields are defined in Section 6.3.2.7 of JSON Web Algorithms */
+        @Serializable
+        data class RsaOtherPrimesInfo(val r: String, val d: String, val t: String)
     }
 }
